% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddtools.R
\docType{data}
\name{indh}
\alias{indh}
\title{INDH data set}
\format{
A data frame with two variables with 720 observations each
}
\source{
Development Aid and social Dyanmics website: \url{https://qua.st/Development-Aid-Social-Dynamics}
}
\description{
Data from the Initiative Nationale du Development Humaine, collected as the part of the SNSF project "Development Aid and Social Dynamics"
}
\examples{
# load the data
data(indh)

# construct rdd_data frame
rdd_dat_indh <- rdd_data(y=choice_pg, x=poverty, data=indh, cutpoint=30)

# inspect data frame
summary(rdd_dat_indh)

# perform non-parametric regression
( reg_np_indh <- rdd_reg_np(rdd_dat_indh) )
}
\references{
Arcand, Rieger, and Nguyen (2015) 'Development Aid and Social Dyanmics Data Set'
}
