% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Lee2008-data.R
\docType{data}
\name{Lee2008}
\alias{Lee2008}
\title{Dataset used in Lee (2008)}
\format{A data frame with 6558 observations and two variables:
\describe{
\item{x}{Vote at election t-1}
\item{y}{Vote at election t}
}}
\source{
Guido Imbens webpage: \url{http://scholar.harvard.edu/imbens/scholar_software/regression-discontinuity}
}
\usage{
Lee2008
}
\description{
U.S. House elections data
}
\examples{
data(Lee2008)
RDDlee <- rdddata(x=x, y=y, data=Lee2008, cutpoint=0)
summary(RDDlee)
plot(RDDlee)
}
\references{
Imbens, Guido and Karthik Kalyanaraman. (2012) "Optimal Bandwidth Choice for the regression discontinuity estimator,"
Review of Economic Studies (2012) 79, 933-959

Lee, D. (2008) Randomized experiments from non-random selection in U.S. House elections,
\emph{Journal of Econometrics}, 142, 675-697
}

