% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reg_np.R
\name{RDDreg_np}
\alias{RDDreg_np}
\title{Parametric polynomial estimator of the regression discontinuity}
\usage{
RDDreg_np(RDDobject, covariates = NULL, bw = RDDbw_IK(RDDobject),
  slope = c("separate", "same"), inference = c("np", "lm"),
  covar.opt = list(slope = c("same", "separate"), bw = NULL))
}
\arguments{
\item{RDDobject}{Object of class RDDdata created by \code{\link{RDDdata}}}

\item{covariates}{TODO}

\item{bw}{A bandwidth to specify the subset on which the parametric regression is estimated}

\item{slope}{Whether slopes should be different on left or right (separate), or the same.}

\item{inference}{Type of inference to conduct: non-parametric one (\code{np}) or standard (\code{lm}). See details.}

\item{covar.opt}{Options for the inclusion of covariates. Way to include covariates, either in the main regression (\code{include}) or as regressors of y in a first step (\code{residual}).}
}
\value{
An object of class RDDreg_np and class lm, with specific print and plot methods
}
\description{
Compute a parametric polynomial regression of the ATE,
possibly on the range specified by bandwidth
}
\examples{
## Step 0: prepare data
data(Lee2008)
Lee2008_rdd <- RDDdata(y=Lee2008$y, x=Lee2008$x, cutpoint=0)
## Step 2: regression
# Simple polynomial of order 1:
reg_nonpara <- RDDreg_np(RDDobject=Lee2008_rdd)
print(reg_nonpara)
plot(reg_nonpara)
}
\references{
TODO
}
\seealso{
\code{\link{RDDbw_IK}} Bandwidth selection using the plug-in bandwidth of Imbens and Kalyanaraman (2012)
}

