% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rddtools.R
\docType{package}
\name{rddtools}
\alias{rddtools}
\alias{rddtools-package}
\title{Regression Discontinuity Design}
\description{
Regression Discontinuity Design
}
\details{
Provides function to do a comprehensive regression discontinuity analysis.
}

