\name{ROT_bw}
\alias{ROT_bw}
\title{Bandwidth selector}
\usage{
ROT_bw(object)
}
\arguments{
  \item{object}{object of class RDDdata}
}
\description{
implements dpill
}
\examples{
#No discontinuity
}
\author{
Drew Dimmery <\email{drewd@nyu.edu}>
}
\references{
McCrary, Justin. (2008) "Manipulation of the running
variable in the regression discontinuity design: A density
test," \emph{Journal of Econometrics}. 142(2): 698-714.
\url{http://dx.doi.org/10.1016/j.jeconom.2007.05.005}
}

