% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_data_methods.R
\name{as.lm}
\alias{as.lm}
\title{Convert a rdd object to lm}
\usage{
as.lm(x)
}
\arguments{
\item{x}{An object to convert to lm}
}
\value{
An object of class \code{lm}
}
\description{
Convert a rdd object to lm
}
\examples{
data(house)
house_rdd <- rdd_data(y=house$y, x=house$x, cutpoint=0)
reg_para <- rdd_reg_lm(rdd_object=house_rdd)
reg_para_lm <- as.lm(reg_para)
reg_para_lm
plot(reg_para_lm, which=4)
}
\seealso{
\code{\link{as.npreg}} which converts \code{rdd_reg} objects into \code{npreg} from package \code{np}.
}
