% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covarTests.R
\name{covarTest_dis}
\alias{covarTest_dis}
\alias{covarTest_dis.RDDdata}
\alias{covarTest_dis.RDDreg}
\title{Testing for balanced covariates: equality of distribution}
\usage{
covarTest_dis(object, bw, exact = NULL, p.adjust = c("none", "holm", "BH",
  "BY", "hochberg", "hommel", "bonferroni"))

\method{covarTest_dis}{RDDdata}(object, bw = NULL, exact = FALSE,
  p.adjust = c("none", "holm", "BH", "BY", "hochberg", "hommel",
  "bonferroni"))

\method{covarTest_dis}{RDDreg}(object, bw = NULL, exact = FALSE,
  p.adjust = c("none", "holm", "BH", "BY", "hochberg", "hommel",
  "bonferroni"))
}
\arguments{
\item{object}{object of class RDDdata}

\item{bw}{a bandwidth}

\item{exact}{Argument of the \code{\link{ks.test}} function: NULL or a logical indicating whether an exact p-value should be computed.}

\item{p.adjust}{Whether to adjust the p-values for multiple testing. Uses the \code{\link{p.adjust}} function}

\item{\ldots}{currently not used}
}
\value{
A data frame  with, for each covariate, the K-S statistic and its p-value.
}
\description{
Tests equality of distribution with a Kolmogorov-Smirnov for each covariates, between the two full groups or around the discontinuity threshold
}
\examples{
data(Lee2008)

## Add randomly generated covariates
set.seed(123)
n_Lee <- nrow(Lee2008)
Z <- data.frame(z1 = rnorm(n_Lee, sd=2),
                z2 = rnorm(n_Lee, mean = ifelse(Lee2008<0, 5, 8)),
                z3 = sample(letters, size = n_Lee, replace = TRUE))
Lee2008_rdd_Z <- RDDdata(y = Lee2008$y, x = Lee2008$x, covar = Z, cutpoint = 0)

## Kolmogorov-Smirnov test of equality in distribution:
covarTest_dis(Lee2008_rdd_Z, bw=0.3)

## Can also use function covarTest_dis() for a t-test for equality of means around cutoff:
covarTest_mean(Lee2008_rdd_Z, bw=0.3)
## covarTest_dis works also on regression outputs (bw will be taken from the model)
reg_nonpara <- RDDreg_np(RDDobject=Lee2008_rdd_Z)
covarTest_dis(reg_nonpara)
}
\author{
Matthieu Stigler <\email{Matthieu.Stigler@gmail.com}>
}
\seealso{
\code{\link{covarTest_mean}} for the t-test of equality of means
}

