% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/covarTests.R
\name{covarTest_mean}
\alias{covarTest_mean}
\alias{covarTest_mean.RDDdata}
\alias{covarTest_mean.RDDreg}
\title{Testing for balanced covariates: equality of means with t-test}
\usage{
covarTest_mean(object, bw = NULL, paired = FALSE, var.equal = FALSE,
  p.adjust = c("none", "holm", "BH", "BY", "hochberg", "hommel",
  "bonferroni"))

\method{covarTest_mean}{RDDdata}(object, bw = NULL, paired = FALSE,
  var.equal = FALSE, p.adjust = c("none", "holm", "BH", "BY", "hochberg",
  "hommel", "bonferroni"))

\method{covarTest_mean}{RDDreg}(object, bw = NULL, paired = FALSE,
  var.equal = FALSE, p.adjust = c("none", "holm", "BH", "BY", "hochberg",
  "hommel", "bonferroni"))
}
\arguments{
\item{object}{object of class RDDdata}

\item{bw}{a bandwidth}

\item{paired}{Argument of the \code{\link{t.test}} function: logical indicating whether you want paired t-tests.}

\item{var.equal}{Argument of the \code{\link{t.test}} function:  logical variable indicating whether to treat the two variances as being equal}

\item{p.adjust}{Whether to adjust the p-values for multiple testing. Uses the \code{\link{p.adjust}} function}

\item{\ldots}{currently not used}
}
\value{
A data frame with, for each covariate, the mean on each size, the difference, t-stat and ts p-value.
}
\description{
Tests equality of means by a t-test for each covariate, between the two full groups or around the discontinuity threshold
}
\examples{
data(Lee2008)

## Add randomly generated covariates
set.seed(123)
n_Lee <- nrow(Lee2008)
Z <- data.frame(z1 = rnorm(n_Lee, sd=2),
                z2 = rnorm(n_Lee, mean = ifelse(Lee2008<0, 5, 8)),
                z3 = sample(letters, size = n_Lee, replace = TRUE))
Lee2008_rdd_Z <- RDDdata(y = Lee2008$y, x = Lee2008$x, covar = Z, cutpoint = 0)

## test for equality of means around cutoff:
covarTest_mean(Lee2008_rdd_Z, bw=0.3)

## Can also use function covarTest_dis() for Kolmogorov-Smirnov test:
covarTest_dis(Lee2008_rdd_Z, bw=0.3)

## covarTest_mean works also on regression outputs (bw will be taken from the model)
reg_nonpara <- RDDreg_np(RDDobject=Lee2008_rdd_Z)
covarTest_mean(reg_nonpara)
}
\author{
Matthieu Stigler <\email{Matthieu.Stigler@gmail.com}>
}
\seealso{
\code{\link{covarTest_dis}} for the Kolmogorov-Smirnov test of equality of distribution
}

