% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dens_test.R
\name{dens_test}
\alias{dens_test}
\title{Run the McCracy test for manipulation of the forcing variable}
\usage{
dens_test(RDDobject, bin = NULL, bw = NULL, plot = TRUE, ...)
}
\arguments{
\item{RDDobject}{object of class RDDdata}

\item{bin}{Argument of the \code{\link{DCdensity}} function, the binwidth}

\item{bw}{Argument of the \code{\link{DCdensity}} function, the bandwidth}

\item{plot}{Whether to return a plot. Logical, default ot TRUE.}

\item{\ldots}{Further arguments passed to \code{\link[rdd]{DCdensity}}.}
}
\description{
Calls the \code{\link[rdd]{DCdensity}} test from package \code{rdd} on a \code{RDDobject}.
}
\examples{
library(RDDtools)
data(Lee2008)
Lee2008_rdd <- RDDdata(y=Lee2008$y, x=Lee2008$x, cutpoint=0)
dens_test(Lee2008_rdd)
}

