% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddtools.R
\docType{package}
\name{rddtools-package}
\alias{rddtools}
\alias{rddtools-package}
\title{Regression Discontinuity Design}
\description{
Set of functions for Regression Discontinuity Design ('RDD'), for data visualisation, estimation and testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bquast/rddtools}
  \item Report bugs at \url{https://github.com/bquast/rddtools/issues}
}

}
\author{
\strong{Maintainer}: Matthieu Stigler \email{Matthieu.Stigler@gmail.com} (\href{https://orcid.org/0000-0002-6802-4290}{ORCID})

Authors:
\itemize{
  \item Bastiaan Quast \email{bquast@gmail.com} (\href{https://orcid.org/0000-0002-2951-3577}{ORCID})
}

}
