% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RDDtools-package}
\alias{RDDtools}
\alias{RDDtools-package}
\title{Regression Discontinuity Design}
\description{
Regression Discontinuity Design
}
\details{
Provides function to do a comprehensive regression discontinuity analysis.
}
\author{
Matthieu Stigler <\email{Matthieu.Stigler@gmail.com}>
}

