% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddtools.R
\docType{package}
\name{rddtools}
\alias{rddtools}
\title{Regression Discontinuity Design}
\description{
Set of functions for Regression Discontinuity Design ('RDD'), for data visualisation, estimation and testing.
}
