% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_rot.R
\name{rot_bw}
\alias{rot_bw}
\title{Bandwidth selector}
\usage{
rot_bw(object)
}
\arguments{
\item{object}{object of class rdd_data}
}
\description{
implements dpill
}
\examples{
#No discontinuity
}
\references{
McCrary, Justin. (2008) 'Manipulation of the running variable in the regression discontinuity design: A density test,' \emph{Journal of Econometrics}. 142(2): 698-714. \url{http://dx.doi.org/10.1016/j.jeconom.2007.05.005}
}
