\name{vcovCluster}
\alias{vcovCluster}
\title{Cluster Heteroskedasticity-consistent estimation of the covariance matrix.}
\usage{
  vcovCluster(object, clusterVar)
}
\arguments{
  \item{object}{Object of class lm, from which RDDreg also
  inherits.}

  \item{clusterVar}{The variable containing the cluster
  attributions.}
}
\value{
  A matrix containing the covariance matrix estimate.
}
\description{
  Offer a cluster variant of the usual
  Heteroskedasticity-consistent
}

